switch build system to guanlecoja
