+++ .gitignore
@@ -1,4 +1,5 @@
-/bower_components/
+dist
+/libs/
 /node_modules/
 
 .DS_Store
