--- bower.json
@@ -1,17 +1,0 @@
-{
-  "name": "angular-dc",
-  "main": "./dist/angular-dc.js",
-  "version": "0.0.4",
-  "dependencies": {
-    "angular": ">=1.2.0",
-    "lodash": ">=2.4.1",
-    "dcjs": "master",
-    "crossfilter": "~1.3.0"
-  },
-  "devDependencies": {
-    "angular-mocks": "latest"
-  },
-  "resolutions": {
-    "crossfilter": "~1.3.0"
-  }
-}
