+++ gulpfile.js
@@ -1,0 +1,2 @@
+require("guanlecoja")(require("gulp"))
+
