+++ package.json
@@ -23,27 +23,9 @@
     }
   ],
   "devDependencies": {
-    "grunt": "~0.4.1",
-    "grunt-contrib-jshint": "~0.6.4",
-    "grunt-contrib-concat": "~0.3.0",
-    "grunt-contrib-uglify": "~0.2.4",
-    "grunt-contrib-watch": "~0.5.3",
-    "grunt-contrib-clean": "~0.5.0",
-    "grunt-contrib-connect": "~0.5.0",
-    "grunt-contrib-less": "~0.7.0",
-    "connect-livereload": "~0.3.0",
-    "grunt-open": "~0.2.2",
-    "grunt-karma": "~0.6.2",
-    "grunt-ngmin": "0.0.3",
-    "grunt-bump": "0.0.11",
-    "matchdep": "~0.1.2",
-    "karma-script-launcher": "~0.1.0",
-    "karma-chrome-launcher": "~0.1.0",
-    "karma-firefox-launcher": "~0.1.0",
-    "karma-html2js-preprocessor": "~0.1.0",
-    "karma-jasmine": "~0.1.3",
-    "karma-phantomjs-launcher": "~0.1.0",
-    "karma": "~0.10.2",
-    "grunt-contrib-qunit": "~0.3.0"
+    "guanlecoja": "^0.3.1",
+    "gulp": "^3.8.8",
+    "run-sequence": "^0.3.7",
+    "shelljs": "^0.3.0"
   }
 }
